+++ CHANGELOG.md
@@ -10,6 +10,8 @@
  + method `add_comment` to class `Site`;
  + method `add_question` to class `Site`;
  + method `delete_answer` to class `Site`;
+ + method `delete_comment` to class `Site`;
+ + method `delete_question` to class `Site`;
  + method `edit_answer` to class `Site`;
  + method `edit_comment` to class `Site`;
  + method `edit_question` to class `Site`;